:sparkles: : 회원가입 유저 플로우 수정 - 닉네임 입력 -> 회원타입 선택에서 - 회원타입 선택 -> 닉네임 입력
