+++ app/src/main/kotlin/com/hm/picplz/ui/screen/sign_up/sign_up_common/views/SignUpSelectTypeScreen.kt
@@ -145,7 +145,7 @@ fun SignUpSelectTypeScreen(
             ) {
                 CommonBottomButton(
                     text = "다음",
-                    onClick = { viewModel.handleIntent(NavigateToSelected) },
+                    onClick = { viewModel.handleIntent( Navigate("sign-up-nickname")) },
                     enabled = currentState.selectedUserType != null,
                     containerColor = MainThemeColor.Black
                 )
@@ -156,11 +156,11 @@ fun SignUpSelectTypeScreen(
     LaunchedEffect(Unit) {
         viewModel.sideEffect.collectLatest { sideEffect ->
             when (sideEffect) {
-                is SignUpSideEffect.SelectUserTypeScreenSideEffect.NavigateToSelected -> {
-                    mainNavController.navigateWithBundle(sideEffect.destination, sideEffect.user)
+                is SignUpSideEffect.Navigate -> {
+                    signUpCommonNavController.navigate(sideEffect.destination)
                 }
                 is SignUpSideEffect.NavigateToPrev -> {
-                    signUpCommonNavController.popBackStack()
+                    mainNavController.popBackStack()
                 }
                 else -> {}
             }
